Re-jigging things a bit, hiding as much of the couch/pouch stuff away.
