+++ packages/backend-core/db.js
@@ -3,5 +3,4 @@ module.exports = {
   ...require("./src/db/constants"),
   ...require("./src/db"),
   ...require("./src/db/views"),
-  ...require("./src/db/pouch"),
 }
