--- packages/backend-core/src/couch/index.ts
@@ -1,4 +1,0 @@
-export * from "./couch"
-export * from "./pouchLike"
-export * from "./utils"
-export { init } from "./pouchDB"
