+++ packages/backend-core/src/db/couch/index.ts
@@ -1,0 +1,4 @@
+export * from "./connections"
+export * from "./pouchLike"
+export * from "./utils"
+export { init, getPouch, getPouchDB } from "./pouchDB"
