+++ packages/backend-core/src/db/couch/utils.ts
@@ -1,6 +1,6 @@
-import { getCouchInfo } from "./couch"
+import { getCouchInfo } from "./connections"
 import fetch from "node-fetch"
-import { checkSlashesInUrl } from "../helpers"
+import { checkSlashesInUrl } from "../../helpers"
 
 export async function directCouchCall(
   path: string,
