+++ packages/backend-core/src/pkg/db.ts
@@ -3,5 +3,4 @@
 export * from "../db"
 export * from "../db/utils"
 export * from "../db/views"
-export * from "../db/pouch"
 export * from "../db/constants"
