+++ packages/backend-core/src/types.ts
@@ -1,1 +1,1 @@
-export { PouchLike } from "./couch"
+export { PouchLike } from "./db"
